requirements.txt
djangorestframework==3.0
six
